# adapter-access/adapter-openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8088
  description: Generated server url
paths:
  /adapter/product/{id}:
    get:
      tags:
      - adapter-controller
      operationId: getProductById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductProductResponse'
  /adapter/product/pan/{pan}:
    get:
      tags:
      - adapter-controller
      operationId: getProductByPan
      parameters:
      - name: pan
        in: path
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ProductProductResponse'
  /adapter/clients:
    get:
      tags:
      - adapter-controller
      operationId: getAllClients
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientResponse'
  /adapter/client/{id}:
    get:
      tags:
      - adapter-controller
      operationId: getClient
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ClientClientResponse'
  /adapter/client/phone-number:
    get:
      tags:
      - adapter-controller
      operationId: getClientByPhoneNumber
      parameters:
      - name: phoneNumber
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ClientClientResponse'
  /adapter/client/pan:
    get:
      tags:
      - adapter-controller
      operationId: getClient_1
      parameters:
      - name: pan
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ClientClientResponse'
  /adapter/client/fio:
    get:
      tags:
      - adapter-controller
      operationId: getClientByFio
      parameters:
      - name: surname
        in: query
        required: true
        schema:
          type: string
      - name: name
        in: query
        required: true
        schema:
          type: string
      - name: lastname
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientClientResponse'
  /adapter/all/clients:
    get:
      tags:
      - adapter-controller
      operationId: getAllClientsResponse
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClientClientResponse'
  /adapter/admin:
    get:
      tags:
      - adapter-controller
      operationId: getAllAdmins
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminResponse'
  /adapter/admin/{id}:
    get:
      tags:
      - adapter-controller
      operationId: getAdminById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdminResponse'
  /adapter/admin/phone-number:
    get:
      tags:
      - adapter-controller
      operationId: getAdminByPhoneNumber
      parameters:
      - name: phoneNumber
        in: query
        required: true
        schema:
          type: string
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AdminResponse'
components:
  schemas:
    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        accountNumber:
          type: string
        cash:
          type: integer
          format: int32
    ProductClientResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        surname:
          type: string
        name:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        isBlocked:
          type: boolean
    ProductProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        pan:
          type: string
        pin:
          type: string
        cvv:
          type: string
        rbs:
          type: string
        active:
          type: boolean
        inAccess:
          type: boolean
        outAccess:
          type: boolean
        isBlocked:
          type: boolean
        expiredDate:
          type: string
          format: date
        clientResponse:
          $ref: '#/components/schemas/ProductClientResponse'
        accountResponse:
          $ref: '#/components/schemas/AccountResponse'
    ClientResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        surname:
          type: string
        name:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        isBlocked:
          type: boolean
    ClientClientResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        surname:
          type: string
        name:
          type: string
        lastName:
          type: string
        phoneNumber:
          type: string
        isBlocked:
          type: boolean
        products:
          type: array
          items:
            $ref: '#/components/schemas/ClientProductResponse'
    ClientProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        pan:
          type: string
        pin:
          type: string
        cvv:
          type: string
        rbs:
          type: string
        active:
          type: boolean
        inAccess:
          type: boolean
        outAccess:
          type: boolean
        isBlocked:
          type: boolean
        expiredDate:
          type: string
          format: date
        accountResponse:
          $ref: '#/components/schemas/AccountResponse'
    AdminResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        surname:
          type: string
        name:
          type: string
        lastname:
          type: string
        phoneNumber:
          type: string
        post:
          type: string
        registration:
          type: string
        password:
          type: string